0926 boardController commit

+++ src/main/java/com/example/demo/board/web/BoardRestController.java
@@ -3,10 +3,16 @@ package com.example.demo.board.web;
 import com.example.demo.board.entity.BoardEntity;
 import com.example.demo.board.service.BoardService;
 import com.example.demo.board.service.request.BoardInsertCommand;
+import com.example.demo.config.auth.PrincipalDetail;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,7 +24,8 @@ public class BoardRestController {
     BoardService boardService;
 
     @PostMapping("/boardAdd")
-    public String boardAdd(BoardInsertCommand command){
+    public String boardAdd(BoardInsertCommand command , @AuthenticationPrincipal PrincipalDetail principalDetail){
+        command.setUser_num(principalDetail.getUserIdx());
         boardService.boardInsert(command);
         return "/board";
     }
@@ -30,15 +37,31 @@ public class BoardRestController {
 
     @GetMapping("/detailOne/{boardNum}")
     public ModelAndView detailOne(@PathVariable int boardNum){
-        Optional<BoardEntity> boardOne = boardService.findOne(boardNum);
+       BoardEntity boardOne = boardService.findOne(boardNum);
         ModelAndView mav = new ModelAndView("boardDetail");
-        mav.addObject("boardOne", boardOne.get());
+        mav.addObject("boardOne", boardOne);
         return mav;
     }
 
     @GetMapping("/delete/{boardNum}")
-    public void delete(@PathVariable int boardNum){
-        boardService.deleteOne(boardNum);
+    public void delete(@PathVariable int boardNum , @AuthenticationPrincipal PrincipalDetail principalDetail, HttpServletResponse response) throws IOException {
+        String email = principalDetail.getUsername();
+        int userIdx = principalDetail.getUserIdx();
+
+        BoardEntity boardOne = boardService.findOne(boardNum);
+        int dbUserIdx = boardOne.getMember().getUserIdx();
+
+        if(userIdx == dbUserIdx) {
+            System.out.println("삭제성공");
+            boardService.deleteOne(boardNum);
+        }else{
+            System.out.println("삭제실패");
+            response.sendRedirect("/board");
+        }
+
+
+
     }
+
 
 }
